requirements.txt
python-telegram-bot==20.5
requests
python-dotenv
Flask>=2.3
aiohttp>=3.8
